fix recoil root builds
