+++ app/packages/app/.env.development.local
@@ -1,0 +1,1 @@
+VITE_API=http://localhost:5151
