+++ app/packages/app/package.json
@@ -42,7 +42,7 @@
         "react-use-measure": "^2.0.1",
         "react-uuid": "^1.0.2",
         "recharts": "2.0.9",
-        "recoil": "^0.7.2",
+        "recoil": "0.7.2",
         "resize-observer-polyfill": "^1.5.1",
         "styled-components": "^5.3.3",
         "uuid": "^8.3.2",
@@ -66,7 +66,7 @@
         "relay-config": "^12.0.1",
         "rollup-plugin-polyfill-node": "^0.6.2",
         "typescript": "4.2.4",
-        "vite": "2.4.2",
+        "vite": "^2.9.9",
         "vite-plugin-relay": "^1.0.7"
     }
 }
