+++ app/packages/components/src/components/Theme/index.ts
@@ -1,0 +1,1 @@
+export { default, ThemeContext, useTheme } from "./Theme";
