+++ app/packages/components/src/with/index.ts
@@ -1,2 +1,1 @@
 export { default as withErrorBoundary } from "./ErrorBoundary";
-export { useTheme, default as withTheme } from "./Theme";
