+++ app/tsconfig.json
@@ -1,0 +1,10 @@
+{
+    "files": [],
+    "references": [
+        { "path": "packages/app" },
+        { "path": "packages/components" },
+        { "path": "packages/flashlight" },
+        { "path": "packages/looker" },
+        { "path": "packages/utilities" }
+    ]
+}
